+++ app/src/main/java/com/example/covid/MiddleFragment.kt
@@ -30,80 +30,7 @@ class MiddleFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        if (!isNetworkAvailable()) {
-          alertForInternet()
-        }
-        else{
-            Toast.makeText(context, "Good, you are connected to internet", Toast.LENGTH_SHORT).show()
-        }
 
-        centerImage.setOnClickListener {
-            content.startRippleAnimation()
-            fetchData()
-        }
-    }
-
-    private fun fetchData() {
-        GlobalScope.launch {
-            val okHttpClient= OkHttpClient()
-            val request= Request.Builder().url("https://api.covid19india.org/data.json").build()
-            val response = withContext(Dispatchers.IO){okHttpClient.newCall(request).execute()}
-            if (response.isSuccessful) {
-                val gsonBuilder = GsonBuilder()
-                gsonBuilder.setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-                val gson = gsonBuilder.create()
-                val data = gson.fromJson(response.body?.string(), TotalResponse::class.java)
-
-                launch(Dispatchers.Main) {
-                    //Find a new Way for this to happen.
-                    //Use Bundle and pass by Fragment Manager
-                   /* val intent = Intent(activity, MiddleFragment::class.java)
-                    intent.putExtra("DataAsJson", gson.toJson(data))
-                    startActivity(intent)  */
-                }
-            }
-        }
-    }
-
-    private fun isNetworkAvailable(): Boolean {
-        val cm = activity?.getSystemService(CONNECTIVITY_SERVICE) as ConnectivityManager
-        val activeNetwork = cm.activeNetworkInfo
-        if (activeNetwork != null) {
-            // connected to the internet
-            if (activeNetwork.type == ConnectivityManager.TYPE_WIFI) {
-                return true
-            } else if (activeNetwork.type == ConnectivityManager.TYPE_MOBILE) {
-                return true
-            }
-            return true
-        }
-        return false
-    }
-
-    private fun alertForInternet() {
-        lateinit var alertDialog: AlertDialog
-        val builder = AlertDialog.Builder(activity)
-        builder.setTitle("No Internet!!")
-        builder.setMessage("Turn On Internet!!")
-        builder.setIcon(android.R.drawable.ic_dialog_alert)
-
-        builder.setPositiveButton("Retry") { dialogInterface, which ->
-            if (isNetworkAvailable()) {
-                alertDialog.cancel()
-            }
-            else
-            {
-                alertForInternet()
-            }
-        }
-        builder.setNegativeButton("Quit") { dialogInterface, which ->
-            activity?.finish();
-            System.exit(0);
-        }
-
-        alertDialog = builder.create()
-        alertDialog.setCancelable(false)
-        alertDialog.show()
     }
 
 }
