+++ app/src/main/res/layout/fragment_middle.xml
@@ -1,22 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.skyfishjy.library.RippleBackground xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".MiddleFragment"
-    android:id="@+id/content"
-    app:rb_color="#0099CC"
-    app:rb_radius="32dp"
-    app:rb_rippleAmount="4"
-    app:rb_duration="3000"
-    app:rb_scale="6">
+    tools:context=".MiddleFragment">
 
-    <ImageView
-        android:layout_width="64dp"
-        android:layout_height="64dp"
-        android:layout_centerInParent="true"
-        android:id="@+id/centerImage"
-        android:src="@drawable/bacteria"/>
 
-</com.skyfishjy.library.RippleBackground>
+</androidx.constraintlayout.widget.ConstraintLayout>
